Fix flakiness in lifecycle tests
Summary: There is some flakyness in the lifecycle tests from cases
where the timing for two actions happen at the same time, and it's not
guaranteed that one will be executed before the other.

This diff changes that so that actions will not end on the same 10ms
time period, they will be spaced out by at least 10ms from each other.

Test Plan: ran the test 1000 times and there were no errors.