thrift: Disable buffer pooling (#861)
This stops the Thrift encoding from pooling buffers because currently
the buffers are returned to the pool before their lifetime is over.
This is because lists, sets and maps in ThriftRW's `wire.Value`s are
lazy; they don't attempt to deserialize until the contents are
requested, which happens when the generic `wire.Value` representation is
converted into generated types using `FromWire`.

To help visualize this,

    // YARPC defines a function that returns a wire.Value, using a
    // pooled buffer during deserialization.
    func Call() wire.Value {
        buf := buffer.Get()
        defer buffer.Put(buf)

        value := decode(buf)
        return value
    }

    // The generated client calls into YARPC to get the decoded
    // wire.Value. It then deserializes the wire.Value into the
    // generated MyType.
    func MyCall() MyType {
        value := yarpc.Call()

        var x MyType
        x.FromWire(value)  -----
        return x
    }

    // MyType contains a list of some other type which it reads from the
    // wire.Value lazy list.
    FromWire(v wire.Value) {
        v.GetList().ForEach(func(item wire.Value) {
            appendToSelf(...)
        })
    }

    // The lazy list now attempts to decode the items from the list into
    // wire.Values on demand, starting at the position it previously
    // recorded.
    func ForEach(f func(wire.Value)) {
        for i := 0; i < length; i++ {
            value := decodeAt(startingPosition)
            f(value)
        }
    }

By the time we get to FromWire, we have already returned the buffer to
the pool so anything ForEach reads from the buffer is now wrong.