Fix flaky Redis test (#858)
The current inbound code makes it very, very difficult to make robust assertions
about the number of messages handled, since any such assertions depend on the
goroutine scheduler.

Instead, convert the test to use the unexported `handle` API directly. We'll
need a better approach before the Redis transport is GA, but this is fine for
now.

Fixes #831.