InjectClients: Support Thrift options
This adds support for specifying Thrift options `multiplexed` and
`enveloped` for clients instantiated using `yarpc.InjectClients`. This
is done by adding a `thrift` tag:

    type handler struct {
      Client keyvalueclient.Interface `service:"foo" thrift:"enveloped"`
    }

Now,

    var h handler
    yarpc.InjectClients(dispatcher, &h)

Is equivalent to,

    h := handler{
      Client: keyvalueclient.New(
          dispatcher.ClientConfig("foo"), thrift.Enveloped)
    }

We do this by altering the RegisterClientBuilder API slightly: In
addition to accepting functions with the signature

    func(transport.ClientConfig) T

Where T is the client interface type, the RegisterClientBuilder now also
accepts functions with the signature,

    func(transport.ClientConfig, reflect.StructField) T

The second parameter provides access to the struct tags, etc. for the
field being filled with that client.

Resolves #772