internal/decode: Temporarily vendor mapdecode (#936)
In the process of working on config interpolation, we are making changes
to internal/decode which we want to eventually push to mapdecode. Since
mapdecode has already diverged from internal/decode, we're temporarily
vendoring mapdecode into internal/ and iterating on that.

Once the library reaches an agreeable state, we can copy the changes
back to mapdecode.