Add logging to outbound unary calls
This PR addresses the next step of T841919.

It takes the existing logging middleware, which only works for unary inbound
calls, and extends it to also work for unary outbound calls. This required a
little refactoring and renaming, since I didn't initially realize that the
method names for inbound and outbound middlewares don't clash.