pally: Add histograms for tracking latency (#937)
To round out the basic metric types, we need some way to track RPC latencies.
The best approach is to use histograms - it's both more correct and matches the
stats that Muttley collects.

This PR adds scalar and vectorized histograms to Pally, along with test
coverage. (See the package-level Pally documentation for an explanation of
vector types).

Down the road, we'll need to come back and optimize the Tally export path; right
now, we take the simple approach of just calling
`tally.Histogram.ObserveDuration` multiple times with the same value. This isn't
ideal, since it forces Tally to binary search for the correct histogram bucket
repeatedly.

This is one sub-task of T841911.