http/out: Support AddHeader option
This adds an `AddHeader` option to HTTP outbounds. This accounts for use
cases where you're trying to call an existing service via YARPC where
the service expects a specific static header to always be present in
outgoing requests.

Note that this feature is intentionally very limited: We don't want to
leak transport-specific details at call sites so we require that the
headers are static and that they are specified at transport construction
time.

Related: T865043