config: Support the Cherami transport (#869)
This adds support for the Cherami transport to the configuration system
defined in #747.

The InboundConfig and OutboundConfig structs used by the Cherami
NewInbound and NewOutbound constructors were renamed to InboundOptions
and OutboundOptions to free up the *Config names for use by the
configuration system. This was to keep the naming consistent with the
HTTP (#748), Redis (#749), and TChannel (#750) configuration systems.