thriftrw-plugin: Test coverage
We were already testing the ThriftRW plugin but not counting it as
covered because of how it is executed. The test runs an external
process (`thriftrw`) which then runs the plugin process
(`thriftrw-plugin-yarpc`), which contains the code being tested.

This changes execution so that `thriftrw` is actually calling a script
which *pretends* to be the ThriftRW plugin but really just forwards
input and output back to our test. The test acts as the plugin, giving
us a better view into our plugin's coverage.