codecov: Set range to 50-90 (#986)
This makes it so that anything at 50% coverage or lower shows up as red
and anything 90% or higher shows up as green. Since we're aiming for 90
to begin with, this will help find low hanging fruits much better than
the default range of 70-100. We can readjust this once we hit 90.