config/tchannel: Improve configuration format (#993)
This changes the TChannel configuration to be more usable. Previously,
the address to listen on was provided in the `transports` section
because that's how the TChannel transport works. This meant there was
nothing to configure for the TChannel inbound and you'd have to add a
`tchannel: {}`  entry to it.

    inbounds:
      tchannel: {}
    transports:
      tchannel:
        address: :4040

With this change, we're changing the format to accept the address in the
inbounds section, making the transport section empty/irrelevant for now.

    inbounds:
      tchannel:
        address: :4040

As a result of this, at most one TChannel inbound may be specified for a
service.