config/chooser: Explanatory comments for peer list (#999)
This adds some comments to the code for building peer lists so that it's
simpler to understand how this relates to the actual config being
decoded.

In the process, I also added a sanity check to the peer list updater
decoding logic: We ensure that *exactly* on peer list updater was
specified rather than going with the first one we find. Otherwise, we
may have a situation where two peer list updaters were written in the
config and the non-deterministic map iteration picks a random one each
time.