chown the mounted volume after generation (#990)
For `make generate` with Docker we mount the local workdir withing the
container. Any files written from withing the container will be owned by
whatever user is loggued in (alwaysr root in our case). The container is
already root on the workdir. We can the chown recursively the workdir to the
owner of the workdir.

The chown only runs withing the docker container. For this to work, a new env
var `WITHIN_DOCKER` is passed to the container.

Note: for what I know. On Mac, you run a linux virtual machine, which in turns
runs containers. You are loggued as root on the linux VM, so you have no
resstrictions. And the virtual machine shares the filesystem between Mac and
the linux VM by rewriting on the fly permissions from the owner of the VM to
root.