Support Non-Running Updates to the RoundRobin Peer List
Summary: By default, the RoundRobin peer list could only accept updates
while the peer list was "Running".  If it got an update request while it
was not running, the peer list would block that update thread until the
peer list started, or a timeout was reached.

This diff alters the RoundRobin list such that it can support taking
updates to the list before it has been started, and after it has been
stopped.  This is done through an "UninitializedPeerList" that accepts
all updates while the peer list is not started.

On Start/Stop the peers are transfered between the Uninitialized peer
list and the Available/Unavailable lists.

**Additionally** this change made the StartupWait peerlist option obsolete,
so I removed it.

Test Plan: Added more test cases for the round robin to validate updates
before and after start/stop. And updated old tests.