Move RoundRobin PeerList out of /x/
Summary: After #1005 the round robin peer list API
is ready to go 1.0.  We may make changes to the internals of the peer
list, but we will not change the API surface of the Config Spec, and the
New function (with list options).

In the process of this diff I removed the api surface of the
PeerRing because that should have never been public, luckily no one
uses it (according to code sleuthing).

I'm leaving behind a "Deprecated" version of the New and Capacity
functions since those are the only functions that are being used.
Once we update the users, we can remove the deprecated functions.

Test Plan: Tests still pass, wrote a deprecated create function.

Depends on: #1005