Fix race conditions in hostport Peer
Summary: In some internal testing we noticed that go test -race was showing race
conditions in the code.  With closer inspection we found that the hostport subscribers
and connectionState were not thread-safe.  This PR adds a read-write mutex around the
peer's subscriber list, and wraps the connectionState in an atomic.

Test Plan: Wrote a concurrent test that failed with race conditions, the change removed
the race condition failures