config: Rename PeerList to PeerChooser
This renames the config.PeerList struct which users embed in their
outbounds to config.PeerChooser since that's what it actually
represents. As a result of this, the BuildPeerList method on it was
renamed to BuildPeerChooser and the PeerListPreset struct (and the
corresponding attribute in TransportSpec) was renamed to
PeerChooserPreset(s).