cleanup: Will not revisit TChannel service name
Having evaluated this note to self, we should try to relax the need for
specifying the service name on the TChannel transport.
TChannel needed this value to construct a sub channel, but we are using
the root peer list and a root handler directly now.
Consequently, we can eventually relax the error that TChannel's
constructor emits if the service name is missing, and will not need to
reintroduce the same error here.