Move observer outbound middleware to the end of the middleware chain
Summary: Currently if you define outbound middleware, it will get
applied after the observer middleware is applied.  Which means the
observer middleware does not actually represent how many requests were
sent.  This PR moves the observer middleware to be the last middleware
in the chain that will be applied immediately before the outbound call.