Fix /procedures docs and provide clarity around CLI options (#1066)
This changes the docs to fix `yarpc/procedures` to `yarpc::procedures`. At the same time, it describes how to use that package (which was unclear) and makes a note that it is experimental. It also adds a section on how to register a JSON handler and mentions yab which was previously more difficult to discover.