[retry middleware][Part 2] ReReader to read the same io.Reader multiple times
Summary: In order to create a retry middleware, we need to be able to
reuse the request's io.Reader body for each retry.  In order to do this
I'm introducing a ReReader that will take an io.Reader and will copy
it's value into a buffer so it can be reused for multiple reads. In
between attempts we will reset the ReReader so that it will begin the
next read cycle at the beginning of the request's body.

Test Plan: Added a series of tests of the ReReader.

Known limitations:  I've added a todo section but, in general:
- We should optimize by casting the source buffer into a bytes buffer
  and using that instead of one of our pooled buffers.
- Will not work if the reader is not exausted to EOF before resetting
  there are some workaround for this if it becomes an issue.
- Does not work concurrently, which should be fine for retries, but will
  have limitations if we want to use it for fanout.