[config] Interpolate values for Peer Choosers and Peer Updaters
Summary: Discovered that the "interpolate" option in config did not
function for attributes on peer chooser and peer updater configs.  This
pr updates the config code such that interpolation resolver is stored on
the kit, and thus usable when decoding the chooser and updater.

Test Plan: Added the test section first, it failed, then I updated the
code to pass the tests.