ThriftRW plugin: Generate ${service}fx module
This changes the ThriftRW plugin for YARPC to generate an Fx-compatible
module for all services. This module currently exports a function that
returns a client constructor. Usage will be as follows,

    fx.Provide(myservicefx.Client("myservice"))

This makes it possible for Fx users to inject clients into the container
without interacting directly with `yarpc.Dispatcher`.