[Retry Middleware][Part 10.5] Prepare everything to be in an /x/ package (#1184)
Summary: With some of the subsequent PRs on the Retry Middleware stack,
it was found to be constricting that we could only create middleware
directly from config.  In order to expose completely customizable retry
middleware, we will be moving basically all this code into an
experimental folder.  This PR in particular goes through all the
existing code and cleans up anything that will get exposed.  In
particular:

- Expose the "ProcedurePolicyProvider" so that it can be constructed
externally.
- Cleanup comments.
- Enforce a solid Options pattern like akshay described in #1164
- Some other minor twinges.

I'm making another easy PR just for the switch to the x folder, happy to
put that in this PR if people protest.