[Retry Middleware][Part 11] Config for creating retry PolicyProviders (#1156)
Summary: This PR introduces a config that creates a retry.PolicyProvider
which can be injected into a RetryMiddleware. In the process I changed
the PolicyProvider from a function to an interface (all internal so it's
nbd, but it was easier to assert on the types for test validation).

Up to this point all the API exposure has been internal, and this is
intentional.  The next PR will introduce an experimental "retry" folder
for creating retry middleware that uses all these objects.

Test Plan: Added config tests.  Needed to add a test func