Add more retryable codes
Summary: Adds more codes of errors that can be retried.  Unavailable is
pretty straightforward, and I'm including Unknown provisionally.
In the process of determining that I needed this, I also found that
http outbounds were not wrapping connection errors with a yarpc type, so
I'm going with "unknown" for now.