Add more retryable codes (#1198)
Summary: Adds more codes of errors that can be retried with idempotent endpoints.  Unavailable is
pretty straightforward, and I'm including Unknown provisionally.
In the process of determining that I needed this, I also found that
http outbounds were not wrapping connection errors with a yarpc type, so
I'm going with "unknown" for now.

In the future, the outbounds should handle the unavailable errors retries internally (so that they can be applied to all endpoints (not just idempotent ones).