Solidify ReReader so we can Reset unread and partially read readers (#1199)
Summary: The Retry Middleware uses a ReReader to read the same source
reader multiple times.  On the first read, it will copy the contents of
the source into a bytes buffer so it can reuse it later. Initially, this
design did not handle what to do if the rereader was reset before the
body had finished reading.  This PR changes the logic so that one of
three things happens:
1) If we haven't read any of the src yet, we don't need to reset
anything, we'll read it in the next run.
2) If we have partially read the src, we'll forcibly read the rest into
the buffer and the next iteration will read from the buffer.
3) If we have finished reading the src, we'll read from the buffer on
the next iteration.

The PR itself removes the TeeReader since we want to use "ReadFrom" to
finish reading the src into our buffer.

In the process I also accounted for the cases where the src is a *bytes.Buffer, and the case where it's already a *Rereader.

Test Plan: Added more tests.