Get cherami and redis to be part of golang crossdock testing (#1219)
This PR sets up a new make target, compose-codecov, that will run all tests within a Docker container using docker-compose that is linked to running Cherami and Redis containers. Two environment variables, CHERAMI=enabled and REDIS=enabled, are also set on the test container, which now indicates to the tests in internal/crossdock to add cherami,redis as oneway transports to test. The net effect is that golang crossdock testing can now test the Cherami and Redis transports. The net effect for code coverage is to increase code coverage by about 1.5%. There is no effect on any other existing part of the setup, minus a few small changes to how code coverage is run, and some bug fixes.

I've also removed the split between experimental and main for Codecov - this PR brings the difference to a much smaller amount, and having both values displayed is a little confusing, and doesn't reflect the number we really track which is the overall code coverage percentage. The overall code coverage percentage is also what is displayed on the main Codecov page for yarpc-go and on the readme badge. I think that if code is worth being in yarpc-go, it should be equally tested.

One note is that the time that code coverage takes in Travis CI is significantly increased for a few reasons, however we can optimize this later.