Cleanup Travis CI (#1229)
This PR does a few things:

Stop using the cache. We do everything in Docker now, and the only thing cached is the Docker binary, which is almost quicker to download than loading and saving the cache.
Set SUPPRESS_CROSSDOCK on all tests except crossdock, so that the travis-docker-push.sh script isn't even invoked.
Stop caching Docker images. Travis actually recommends not doing this, there was barely any speedup (it might actually be slower) and the images were getting rebuilt anyways, and there were caching errors on builds.