glide up and regenerate ThriftRW code (#1248)
This unpins gomock in testImport so that we're okay with any version of
gomock. Some dependency of ours (I can't tell which; ThriftRW was one of
them but that was fixed in v1.5.0) pins gomock in its `import` section
rather than `testImport`.

Because the version of ThriftRW was updated, we also need to regenerate
all ThriftRW code.