Add yarpc error code information to peer lists (#1263)
Summary: Currently if a yarpc peer list (round robin or peer heap) has
an error when choosing a peer, we return a "raw" error.  The rest of the
yarpc stack should not have to worry about the errors returned from the
peer chooser, the chooser should wrap the errors it returns in yarpc
error codes for the different possible issues it could experience.  This
PR wraps the roundrobin and peer heap implementations in yarpc errors
for the "Choose" call stack.

Test Plan: Updated tests, they pass.