[RetryStats][Part 1] Move observability.digester into reusable package
Summary: I'm going through and adding better metrics to the retry
middleware stack. This involves taking as much as possible from the
current observability middleware and using it for other purposes.

This PR takes the digester out of the observability folder so it can be
reused in the retry metrics (and beyond).

Test Plan: tests pass