Read entire resp body into buffer for retries (#1294)
Summary: Depending on the transport, we might have cases where
the "success" of a request is not determined until the io.ReadCloser
body of the Response has been fully read (example being tchannel
requests that have more than 1 frame (1<<16 bytes)).  If we pass
back the body to the handler and reading the body fails, we no longer
have the ability to retry the request.

This PR reads the response body into a "buffercloser" type that is
a pooled buffer that will return itself to the sync pool when the
"Close" method is called. It also supports "ReadFrom" for wrapping
the incoming io.Reader, and io.WriteTo to allow io.Copy without
allocations.

Test Plan: added tests, have tested this in the frontcar repo with
its megasuite integration tests.