Pull backoff.None out of the internal backoff package. (#1302)
Summary: I'm going through the internal dependencies of the retry module
so I can pull retry middleware into it's own repo (Pinning to yarpc
versions is a bit painful when not much is changing (though we haven't
fully committed to the api.

This is a small dependency on the backoff interface, I could copy the
code if I wanted, but it's not going to be changing anytime soon since
it's already a part of the api.