Undeprecate typed yarpcerror creation and validation functions (#1312)
Summary: #1303 ended up adding new yarpc error apis and deprecating
these convenience functions for creating errors.  I think the overall
direction of #1303 is better in general but I think the convenience of
these functions should not be understated.  Instead of writing

  yarpcerrors.Newf(yarpcerrors.CodeInternal, "test")

I want to write:

  yarpcerrors.InternalErrorf("test")

Which <in my opinion> is a much cleaner api.

Full disclosure, I have a lot of internal packages that are using this
API in tests/etc.  And internally, using a deprecated API will cause
your lint to fail... so there's that.

In terms of what we encourage customers to use, I'm pretty flexible, but
I think the api has a lot of value, and it's frustrating to be
discouraged from using it.