Remove x/retry
Summary: It's become a difficult challenge iterating fast on the retry
middleware stack, with plans for an open source pally, we decided to
iterate on the apis for the retry middleware internally before we make
an external version of the retry middleware.  Once we do we'll probably
skip the /x/ package and go straight to a non-x folder.