[RFC] Bypass "ClientConfig" with new hotness "OutboundConfig"
Summary: The ClientConfig interface we've been working with has some
drawbacks.  Primarily, it panics EVERYWHERE (*mostly everywhere). So
as we start thinking about dynamic config for things using YARPC, we
start to get into a dangerous zone where all our functions are incredibly
panic-prone. So if we wanted to check if an outbound key existed, the
ClientConfig apis would probably panic.

Additionally, the fact that both Unary and Oneway are hardcoded into the
ClientConfig interface means that we cannot iterate with new Outbound
types (*cough* streaming *cough*).

This PR provides an alternative to ClientConfig, "OutboundConfig".  The
OutboundConfig implements the ClientConfig interface (and functionality)
for backwards compatibility, but, it exposes all of the struct
properties externally, so if we want to add more types to the Outbounds
struct, we can. It also provides two functions for forcibly getting
(panic) and optionally getting an OutboundConfig.