[AppErr][part 1] Add "Full" Application Error to transport.Response
Summary: We want to have the ability in middleware to introspect
application error codes, this PR adds a full application error
to the transport.Response in order to introspect that info (though
encodings/middleware/transports need to read/write to it).

It also adds a new "ResponseWriterWithResponse" which is a bit of a
round-about step 1 to potentially deprecate the ResponseWriter API.
Essentially, in fear of having to make more API changes here (if we want
to pass more information in the ResponseWriter),  we introduce a struct that
can be extended with more response information, but, basically, the
transport.Response already has all the attributes we might ever need, so
we just create a *Response with the ResponseWriter and allow alteration
of these Response attributes directly through this new API.

This is flexible/we can change this, feel free to add your thoughts.

The next PR does a refactor of the http handler to facilitate the new
ResponseWriterWithResponse interface, it might be easier to get an idea
about the changes involved there.