[AppErr][Part 2] Refactor http handler
Summary: In order to slide in the AppErr negotiation and http support
for the new ResponseWriterWithResponse interface, this PR refactors the
yarpc http handler to use a transport.Response inside the
ResponseWriter.  In the process of doing this, the core components of
the existing ResponseWriter were pulled out into functions in the main
code path of the http handler. This means that middleware in the inbound
stack will be able to grab the *Response by upcasting the
ResponseWriter, and the logic for adding AppErr negotiation for the wire
will be an added `if` statement in the ServeHTTP function.