[AppErr][Part 3] HTTP Response+Error negotiation
Summary: Now that we've added the new FullAppErr entry points and
refactored the http handler, this PR adds header negotiation for whether
an http response will encode the AppError's error code on the wire (this
includes setting an appropriate http status code).

This is a bit of a scary change, so please lay in the feedback. One
glaring issue is that we aren't propagating the error messages.  Might
be nice to add another header that writes the error message into a
response header (as mentioned in one of peter's comments).

Test Plan: Added tests for both the inbound and outbound case that would
perform the negotiation with and without the proper headers.