[WIP] http client interceptor.
Summary: This is just a test bed I'm playing around with to see if we
can get 100% rest compatibility in yarpc by wrapping requests before
they are sent in the wire.  We're already wrapping all the inbound
requests through the handler API, this follows a similar functionality
except for outbound calls.