Add x/yarpctest folder for integration test infrastructure.
Summary: This is an adaptation of an internal test library we've been
using in the frontcar, the priority of the library being the legibility
of the tests that are written.
Before we add the GRPC streaming functionality in the next PRs, this PR
adds the basic test infrastructure in an /x/ package that can be used to
start up services (lifecycles) and apply requests (actions) against those
services.
So if we want to define a service like this:

```
HTTPService(
  Name("myservice"),
  Port(12345),
  Proc(Name("echo"), EchoHandler()),
)
```

We can make requests against it like this:

```
HTTPRequest(
  Port(12345),
  Body("test body"),
  Service("myservice"),
  Procedure("echo"),
  ExpectRespBody("test body"),
),
```

It's possible that the apis here might need to change based on new
functionality requirements, which is a reason to incubate this in an /x/
package.