[Protobuf Streaming] Add base APIs for Proto Streaming.
Summary: This PRs adds the basic hookins for Protobuf that will allow
adding streaming to the generated protobuf code.  The base components
are:

- Allow extracting streaming outbounds from clientconfigs (this may
introduce a breaking change if people wrap clientconfigs they pass into
protobuf, open to alternatives).
- Add Streaming CallStream function to proto clients.
- Add ability to inject Stream Procedures/Handlers into the proto
BuildProcedures function.
- Add ToReader and ToProtoMessage functions to convert between
proto.Message and ReadClosers, these are convenience functions that should
be used whenever a stream receives or sends a message.