Undeprecate ClientConfig function (#1371)
Summary: Didn't mean to deprecate the Dispatcher's ClientConfig function
yet.  This was meant to be deprecated some time in the future after
we've migrated most of the users off the function.  Currently, any use
of this function will cause lint errors internally, which is a lot of
churn for users.