[Patch] Updated API, Nits, and tests
Summary: This takes the series of feedback on the PR and cleans it up.

We've updated the API to be more open, we pass contexts in most of the
functions, and clarified a bunch of the comments around API areas (and
where certain things apply).  Streaming APIs are more disjoin currently,
so our API will need to be open enough to accomodate different use
cases.

Also cleaned up some code (removed introspection stuff for now) and
added more tests, coverage should go up a lot