[Patch] Update to new APIs, address feedback.
Summary: This commit is the "upgrade" that happened after the api changes in
the previous PRs, the external recv/send APIs of the protobuf handlers haven't
changed, but we'll have the option in the future to add functions that
include a context (for per request timeouts), but we don't need to worry
about that right now.

Additionally, this commit cleans up the ToReader/ToProtoMessage
functions and replaces them with ReadFromStream and WriteToStream (which
significantly cleans up the generated code).