Revert "Add middleware to check if dispatcher is running when making calls and handling requests (#1375)" (#1388)
Summary: Noticed this in a couple internal tests.  If we're explicitly erroring all requests
until the dispatcher is in the "running" mode, we'll break any safe shutdown functionality we have, 
causing a bunch of errors on the edge cases of the dispatcher.  For the case of the outbound ctxErr
we can check the IsRunning function on the chooser to determine the error returned.

The second part of the reverted diff removed http peers from availability if the error returned from
the server was "unavailable", unfortunately, we don't have the ability to distinguish whether this peer
was actually unavailable, or whether the peer was forwarding an unavailable error (for a proxy use case).
In which case, it could cause churn to remove the peer from the availability list (if this were the 
muttley peer this could cause issues).