Add PeerChooserSpec to yarpcconfig
Summary: Was about to make a task for this, but figured it wouldn't take
too long.  We've been getting some requests via sharding and some custom
use cases where "ChooserList" is actually restrictive (we currently
require an Updater whenever you use a chooser).  This is unideal so this
PR introduces a ChooserSpec which will not have the requirement of the
updater.

This results in some overlap between the ListSpec and the ChooserSpec,
for now the model I've gone with is that Choosers take precedence
over Lists if both register under the same name.  We have the option
to fail the registration if there are duplicates. The current List
registration will simply override the existing Spec if it's registered
under the same name, not exactly sure what the "right" thing to do
here is, open to suggestions if people feel differently.

Test Plan: Added tests