WIP: bufferpool: Detect use-after-free / concurrent operations
This is WIP prototype to detect bad use of buffer pools, similar to
https://godoc.org/github.com/m3db/m3x/pool

In production, it will detect concurrent accesses to the same buffer
(Which are usually caused by a use-after-free because a buffer was
released too early). In tests, we can implement additional
use-after-free detection which will disable pooling, but will detect any
operations performed after the free.

WIP WIP WIP: This is not ready for a full code review (E.g., no unit
tests) but to discuss this approach for the buffer pool.